Add files via upload
Script and its function to ensure that points are generated with a minimum point spacing.  The extent and the spacing threshold will determine whether the desired sample size is obtained.  The process essentially throws out batches of points until a complete set of N is obtained rather than trying to incrememtally add points to an existing sample.  This is more 'random' given the restriction of minimum spacing.